requirements.txt
streamlit==1.24.0
mysql-connector-python==8.0.33
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.7.0
pandas==2.0.3